feat: text on both platforms uses alternate build of troika for webgpu (NOTE, webgl of alt version broken) They are taking forever to update
