+++ package.json
@@ -102,7 +102,8 @@
     "stats.js": "^0.17.0",
     "suspend-react": "^0.1.3",
     "three-mesh-bvh": "^0.8.3",
-    "troika-three-text": "file:./lib/troika/troika-three-text-0.52.4.tgz",
+    "troika-three-text": "^0.52.4",
+    "troika-three-text-webgpu": "file:./lib/troika/troika-three-text-0.52.4.tgz",
     "tunnel-rat": "^0.1.2",
     "use-sync-external-store": "^1.4.0",
     "utility-types": "^3.11.0",
