+++ src/core/Performance/DetectGPU/DetectGPU.stories.tsx
@@ -3,9 +3,11 @@ import { Vector3 } from 'three'
 import { Meta, StoryObj } from '@storybook/react-vite'
 
 import { Setup } from '@sb/Setup'
+import { PlatformSwitch } from '@sb/components/PlatformSwitch'
 
 import { DetectGPU } from 'drei'
-import { Text } from '../../UI/Text/Text'
+import { Text as LegacyText } from '../../../legacy/UI/Text/Text'
+import { Text as WebGPUText } from '../../../webgpu/UI/Text/Text'
 
 export default {
   title: 'Performance/DetectGPU',
@@ -24,11 +26,15 @@ type Story = StoryObj<typeof DetectGPU>
 function DetectGPUScene(props: React.ComponentProps<typeof DetectGPU>) {
   return (
     <DetectGPU {...props}>
-      {({ device, fps, gpu, isMobile, tier, type }) => (
-        <Text maxWidth={200}>
-          | device {device} fps {fps} | gpu {gpu} isMobile {isMobile?.toString()} | Tier {tier.toString()} Type {type} |
-        </Text>
-      )}
+      {({ device, fps, gpu, isMobile, tier, type }) => {
+        const info = `| device ${device} fps ${fps} | gpu ${gpu} isMobile ${isMobile?.toString()} | Tier ${tier.toString()} Type ${type} |`
+        return (
+          <PlatformSwitch
+            legacy={<LegacyText maxWidth={200}>{info}</LegacyText>}
+            webgpu={<WebGPUText maxWidth={200}>{info}</WebGPUText>}
+          />
+        )
+      }}
     </DetectGPU>
   )
 }
