+++ src/core/UI/index.ts
@@ -5,4 +5,3 @@ export * from './useCursor'
 export * from './useIntersect'
 export * from './ScreenSpace'
 export * from './Select'
-export * from './Text'
