+++ src/legacy/UI/Text/Text.docs.mdx
@@ -1,6 +1,6 @@
 ---
 title: Text
-sourcecode: src/core/UI/Text/Text.tsx
+sourcecode: src/legacy/UI/Text/Text.tsx
 ---
 
 {/* AUTO:badges */}
