+++ src/legacy/UI/Text/index.ts
@@ -1,0 +1,1 @@
+export * from './Text'
