+++ src/legacy/index.ts
@@ -15,3 +15,4 @@ export * from './Geometry'
 export * from './Materials'
 export * from './Staging'
 export * from './Effects'
+export * from './UI'
