+++ src/webgpu/UI/index.ts
@@ -1,1 +1,2 @@
 export * from './Image'
+export * from './Text'
